Handle spurious wakeups in VerifyClass
Spurious wakeups in VerifyClass can cause the class status to be
the same as the old status. Relax the check to be >= to handle this.

Adding a regression test is infeasible since it is a hard to get a
thread in the "somebody else verifying" loop.

Test: test-art-host
Bug: 62912904

(cherry picked from commit 5ef7020490126d868b10e8d40cdd2963c51eaa6c)

Merged-In: I308a395220507a875bb2fbf0800a14fabb434fda

Change-Id: I535f4f487a99c5e49848a3208ccdf646ca8888f5