Remove obsolete code paths from the ARM code generator
After the last changes to the ARM code generator, several code paths
that handle some HCondition corner cases are rarely executed and are,
strictly speaking, unnecessary because the rest of the compiler can do
their job with minimal modifications (and even generate better code),
but have been kept in order to minimize the differences with the
previous ARM code generator. Now that the latter has been removed, the
obsolete code paths can be deleted as well (practically without any
change in behaviour).

Furthermore, this commit contains a preliminary improved fix for the
issue checked by the 657-branches test. The proper fix, however, should
be in the instruction simplifier or another compiler pass before code
generation.

Test: 657-branches
Test: test-art-target
Change-Id: I7d785a1607bc99bff0bfc33050b567a9cf6925c9