Update the context classpath after the dex files are opened.
The classpath stored in the context maybe be out of sync with the list of
opened dex files if any of the dex file is a multidex or cannot be opened.

This CL ensures that OpenDexFiles updates the classpath with the dex file
locations that were open.

The change does not affect the current use of ClassLoaderContext. It is a
preparatory step for b/62269291

Bug: 62269291
Test: m test-art-host-gtest-class_loader_context_test
Change-Id: Ibe675df9b2c06a4df8189f0d72df8b4061603a0e