ahat: Expand test coverage using static heap dumps.
A limitation of the existing ahat test framework was that it was only
easy to test heap dumps generated as part of building the ahat tests.

This change adds static heap dumps from previous versions of Android
to test issues that were specific to previous versions or that are
easier to test with a hard-coded heap dump file.

The test heap dumps are now stored as resources in ahat-test.jar
rather than being passed to ahat-test.jar using system properties.

Fixes a couple of minor bugs encountered when expanding tests.

There are currently two failing tests:
1. Thread roots are not properly marked.
2. Root records not in the default heap are not properly marked.

Test: m ahat-test, from clean build, with new tests added.
Bug: 65356532

Change-Id: I3fa77e2e75d535a1dd68a763c7c45913a9e4074d