Depend on dalvikvm64 directly instead of using ART_HOST_EXECUTABLES
ART_HOST_EXECUTABLES is a list of phony targets.  A file that depends
on a phony target is always considered dirty and is rebuilt on every
build.  Make the test-dump*.hprof targets depend on the installed
location of dalvikvm64, and pass --64 to the art script so that it
calls dalvikvm64 instead of the dalvikvm symlink.

Bug: 68105950
Test: m out/target/common/obj/JAVA_LIBRARIES/ahat-test-dump_intermediates/test-dump-base.hprof
Change-Id: Ibf2dae4bc6113b53db6ebadc93a67eceb55fd77b