Fix default method verifier check for compact dex
Add a feature flag and mark if the compact dex file was
generated from a dex file that supported default methods or not. This
is done to maintain the existing verifier behavior differences for
dex files that do and don't support default methods.

Fixed callers to use a virtual function instead of always checking
the dex file version.

Re-enabled run-test 975.

Bug: 70930171
Bug: 63756964
Test: test-art-host

Change-Id: I46ac5d3cde0d0c9f41cbc68cccaf954b531e0edf