Change from PPID to CWD to distinguish between different trees
Using PPID to distinguish between multiple concurrent trees being
built was causing the whole build-system to be re-scanned each time we
tried to run an art test. By changing it to be based on CWD ckati
should not need to rescan all mk files as much.

This is a hacky workaround, but the only proper fix is stop using a
test runner built on top of make.

Test: mma -j40 test-art-host-gtest-zip_archive_test64 several times
Bug: 71766439
Change-Id: If17221a74b081ebfec8dbf7076b2cd702861eda9